+++ x/emissions/testutil/test_utils.go
@@ -340,6 +340,7 @@ func (s *TestSuite) SetupTest() {
 		accountKeeper,
 		bankKeeper,
 		emissionsKeeper,
+		nil,
 		authtypes.FeeCollectorName,
 	)